Use EditDiaglogFramgent instead of dialog

+++ app/src/main/java/com/paranoid/mao/tsnddemo/ui/manage/ManageViewModel.kt
@@ -1,6 +1,8 @@
 package com.paranoid.mao.tsnddemo.ui.manage
 
 import android.arch.lifecycle.ViewModel
+import android.bluetooth.BluetoothAdapter
+import com.paranoid.mao.tsnddemo.R
 import com.paranoid.mao.tsnddemo.data.DataRepository
 import com.paranoid.mao.tsnddemo.vo.Sensor
 import io.reactivex.subjects.PublishSubject
@@ -8,18 +10,29 @@ import kotlinx.coroutines.experimental.launch
 import java.util.concurrent.TimeUnit
 
 class ManageViewModel(
-        private val dataRepository: DataRepository) : ViewModel() {
+        private val dataRepository: DataRepository
+) : ViewModel() {
 
     val allSensorList = dataRepository.sensorList
     private val clickSubject = PublishSubject.create<Sensor>()
     val editEvent = clickSubject.throttleFirst(1, TimeUnit.SECONDS)
 
-    fun insert(sensor: Sensor) = launch {
-        dataRepository.insert(sensor)
+    fun insert(sensor: Sensor): Int? {
+        return if (BluetoothAdapter.checkBluetoothAddress(sensor.mac)) {
+            launch {
+                dataRepository.insert(sensor)
+            }
+            null
+        } else {
+            R.string.msg_illegal_mac_address
+        }
     }
 
-    fun delete(sensor: Sensor) = launch {
-        dataRepository.delete(sensor)
+    fun delete(sensor: Sensor){
+        if (sensor == Sensor.DUMMY) return
+        launch {
+            dataRepository.delete(sensor)
+        }
     }
 
     fun click(sensor: Sensor) {
